backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
google-generativeai
pydantic
motor
pymongo
python-dotenv
web3
eth-account
